add metadata for osnap page
